change user into profile and add user table / add logout system

+++ server/src/modules/dragon/dragonRepository.ts
@@ -5,9 +5,21 @@ import type { Dragon } from "../../types/types";
 class dragonRepository {
   // C of CRUD
   async create(dragon: Omit<Dragon, "id" | "strength" | "speed" | "stamina">) {
+    const [profileCheck] = await databaseClient.query<Rows>(
+      `
+      SELECT id
+      FROM profile
+      WHERE user_id = ? AND id = ?
+      `,
+      [dragon.user_id, dragon.profile_id],
+    );
+
+    if (profileCheck.length === 0)
+      throw new Error("Profil non lié à cet utilisateur.");
+
     const [result] = await databaseClient.execute<Result>(
       `
-      INSERT INTO dragon (name, specie_id, user_id, strength, speed, stamina)
+      INSERT INTO dragon (name, specie_id, profile_id, strength, speed, stamina)
       VALUES (?, ?, ?,
         (SELECT base_strength from specie WHERE specie.id = ?),
         (SELECT base_speed from specie WHERE specie.id = ?),
@@ -16,7 +28,7 @@ class dragonRepository {
       [
         dragon.name,
         dragon.specie_id,
-        dragon.user_id,
+        dragon.profile_id,
         dragon.specie_id,
         dragon.specie_id,
         dragon.specie_id,
@@ -29,12 +41,24 @@ class dragonRepository {
   }
 
   // R of CRUD
-  async readAll() {
+  async readAll({ userId, profileId }: { userId: number; profileId: number }) {
+    const [profileCheck] = await databaseClient.query<Rows>(
+      `
+      SELECT id
+      FROM profile
+      WHERE user_id = ? AND id = ?
+      `,
+      [userId, profileId],
+    );
+
+    if (profileCheck.length === 0)
+      throw new Error("Profil non lié à cet utilisateur.");
+
     const [rows] = await databaseClient.query<Rows>(
       `
       SELECT
         dragon.id as dragon_id,
-        dragon.user_id,
+        dragon.profile_id,
         specie.specie,
         dragon.name,
         dragon.adopted_at,
@@ -45,42 +69,36 @@ class dragonRepository {
         specie.url_adult
       FROM dragon
       INNER JOIN specie ON specie.id = dragon.specie_id
+      WHERE profile_id = ?
       `,
+      [profileId],
     );
 
     return rows;
   }
 
-  async readAllByUser(userId: number) {
-    const [rows] = await databaseClient.query<Rows>(
+  async read({
+    userId,
+    profileId,
+    dragonId,
+  }: { userId: number; profileId: number; dragonId: number }) {
+    const [profileCheck] = await databaseClient.query<Rows>(
       `
-      SELECT
-        dragon.id as dragon_id,
-        dragon.user_id,
-        specie.specie,
-        dragon.name,
-        dragon.adopted_at,
-        dragon.strength,
-        dragon.speed,
-        dragon.stamina,
-        specie.url_baby,
-        specie.url_adult
-      FROM dragon
-      INNER JOIN specie ON specie.id = dragon.specie_id
-      WHERE dragon.user_id = ?
+      SELECT id
+      FROM profile
+      WHERE user_id = ? AND id = ?
       `,
-      [userId],
+      [userId, profileId],
     );
 
-    return rows;
-  }
+    if (profileCheck.length === 0)
+      throw new Error("Profil non lié à cet utilisateur.");
 
-  async read({ userId, dragonId }: { userId: number; dragonId: number }) {
     const [rows] = await databaseClient.query<Rows>(
       `
       SELECT
         dragon.id as dragon_id,
-        dragon.user_id,
+        dragon.profile_id,
         specie.specie,
         dragon.name,
         dragon.adopted_at,
@@ -91,9 +109,9 @@ class dragonRepository {
         specie.url_adult
       FROM dragon
       INNER JOIN specie ON specie.id = dragon.specie_id
-      WHERE user_id = ? AND dragon.id = ?
+      WHERE profile_id = ? AND dragon.id = ?
       `,
-      [userId, dragonId],
+      [profileId, dragonId],
     );
 
     return rows[0];
@@ -101,18 +119,30 @@ class dragonRepository {
 
   // U of CRUD
   async update(dragon: Omit<Dragon, "specie_id">) {
+    const [profileCheck] = await databaseClient.query<Rows>(
+      `
+      SELECT id
+      FROM profile
+      WHERE user_id = ? AND id = ?
+      `,
+      [dragon.user_id, dragon.profile_id],
+    );
+
+    if (profileCheck.length === 0)
+      throw new Error("Profil non lié à cet utilisateur.");
+
     const [result] = await databaseClient.execute<Result>(
       `
       UPDATE dragon
       SET name = ?, strength = ?, speed = ?, stamina = ?
-      WHERE user_id = ? AND id = ?
+      WHERE profile_id = ? AND id = ?
       `,
       [
         dragon.name,
         dragon.strength,
         dragon.speed,
         dragon.stamina,
-        dragon.user_id,
+        dragon.profile_id,
         dragon.id,
       ],
     );
@@ -121,13 +151,29 @@ class dragonRepository {
   }
 
   // D of CRUD
-  async destroy({ userId, dragonId }: { userId: number; dragonId: number }) {
+  async destroy({
+    userId,
+    profileId,
+    dragonId,
+  }: { userId: number; profileId: number; dragonId: number }) {
+    const [profileCheck] = await databaseClient.query<Rows>(
+      `
+      SELECT id
+      FROM profile
+      WHERE user_id = ? AND id = ?
+      `,
+      [userId, profileId],
+    );
+
+    if (profileCheck.length === 0)
+      throw new Error("Profil non lié à cet utilisateur.");
+
     const [result] = await databaseClient.execute<Result>(
       `
       DELETE FROM dragon
-      WHERE user_id = ? AND id = ?
+      WHERE profile_id = ? AND id = ?
       `,
-      [userId, dragonId],
+      [profileId, dragonId],
     );
 
     return result.affectedRows;
